refactor: move NumberStepper to beta folder
NumberStepper is now exported from @mrshmllw/smores-react/beta instead of the main export
